backend/pyproject.toml
[project]
name = "artifice-backend"
version = "0.2.0"
description = "Backend API for Artifice Toolkit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.1",
    "pillow>=12.0.0",
    "pydantic-settings>=2.12.0",
    "qrcode[pil]>=8.2",
    "qrlyzer>=0.2.6",
    "uvicorn>=0.38.0",
    "zxcvbn>=4.5.0",
]
